Add qp_utils.py for cache decorator, remove pickl in ei_handler.py
